+++ interform/langs/__init__.py
@@ -1,0 +1,1 @@
+""" Subpackage with realisation of main features for different DIF languages """