+++ interform/langs/ini.py
@@ -1,12 +1,17 @@
+"""
+``INI`` language support.
+
+Work in progress.
+"""
 from configparser import ConfigParser
 
-from ..core import BaseInterchange
+from ..core import BaseLang
 
 
-class INI_Format(BaseInterchange):
+class INI_Format(BaseLang):
     # TODO
 
-    def __dict_from_ini(self, config_parser) -> dict:
+    def __dict_from_ini(self, ini_obj) -> dict:
         # TODO: Needs to be tested
         parser = ConfigParser()
         parser.read(self.__parsed_file)