+++ interform/tests/core.py
@@ -1,7 +1,7 @@
 from shutil import rmtree
 from os import path, chdir
 
-from interform.core import BaseInterchange
+from interform.core import BaseLang
 
 
 temp_dir_path = './temp/'
@@ -40,86 +40,86 @@ def change_path_to_testsdir():
     chdir(path.dirname(__file__))
 
 
-class BaseConfigTest():
+class BaseLangTest():
     """
     Base class for all tests.
-    `setup_method()` method should be overwritten with valid config object initialization.
+    `setup_method()` method should be overwritten with valid object initialization.
     """
     def setup(self):
-        self.config = BaseInterchange()
+        self.language_object = BaseLang()
 
     def test_delete_file(self):
         """
         Delete local file and check
         is the functions output works correctly
         """
-        expect_success = self.config.delete_file()
-        expect_fail = not self.config.delete_file()
+        expect_success = self.language_object.delete_file()
+        expect_fail = not self.language_object.delete_file()
 
         assert expect_success and expect_fail
 
     def test_create_file(self):
         """ Create local file """
-        self.config.create_file()
+        self.language_object.create_file()
 
-        assert self.config.is_file_exist()
+        assert self.language_object.is_file_exist()
 
     def test_file_read(self):
-        assert self.config.read_file_as_dict()["root"]["Person_1"]["name"] == default_local_file_dict["root"]["Person_1"]["name"]
+        assert self.language_object.read_file_as_dict()["root"]["Person_1"]["name"] == default_local_file_dict["root"]["Person_1"]["name"]
 
     def test_commit_changes(self):
         person_name = "Jeff"
-        self.config["root"]['Person_1']['name'] = person_name
-        self.config.commit()
+        self.language_object["root"]['Person_1']['name'] = person_name
+        self.language_object.commit()
 
-        read_dict = self.config.read_file_as_dict()
+        read_dict = self.language_object.read_file_as_dict()
 
         assert read_dict["root"]["Person_1"]["name"] == person_name
 
     def test_clear_dict(self):
         """ Clear the `dictionary` attribute """
-        self.config.clear()
+        self.language_object.clear()
 
-        assert len(self.config) == 0
+        assert len(self.language_object) == 0
 
     def test_refresh_safe(self):
         """
         Refresh `dictionary` from local file in safe mode
         and check for nested dictionary existance inside.
         """
-        self.config["root"]['nested_dict_refresh_test'] = {"will_stay": "cia"}
-        self.config.commit()
-        self.config["root"]['nested_dict_refresh_test']["will_stay_too"] = "wait..."
-        self.config.refresh()
+        self.language_object["root"]['nested_dict_refresh_test'] = {"will_stay": "cia"}
+        self.language_object.commit()
+        self.language_object["root"]['nested_dict_refresh_test']["will_stay_too"] = "wait..."
+        self.language_object.refresh()
 
-        assert self.config["root"]['nested_dict_refresh_test'].get("will_stay", None) is not None and \
-               self.config["root"]['nested_dict_refresh_test'].get("will_stay_too", None) is not None
+        assert self.language_object["root"]['nested_dict_refresh_test'].get("will_stay", None) is not None and \
+               self.language_object["root"]['nested_dict_refresh_test'].get("will_stay_too", None) is not None
 
     def test_refresh_unsafe(self):
-        self.config["root"]['nested_dict_refresh_test'] = {"will_stay": "cia"}
-        self.config.commit()
-        self.config["root"]['nested_dict_refresh_test']["will_be_lost"] = "what!?"
-        self.config.refresh(safe_mode=False)
+        self.language_object["root"]['nested_dict_refresh_test'] = {"will_stay": "cia"}
+        self.language_object.commit()
+        self.language_object["root"]['nested_dict_refresh_test']["will_be_lost"] = "what!?"
+        self.language_object.refresh(safe_mode=False)
 
-        assert self.config["root"]['nested_dict_refresh_test'].get("will_be_lost", None) is None and \
-               self.config["root"]['nested_dict_refresh_test'].get("will_stay", None) is not None
+        assert self.language_object["root"]['nested_dict_refresh_test'].get("will_be_lost", None) is None and \
+               self.language_object["root"]['nested_dict_refresh_test'].get("will_stay", None) is not None
 
     def test_reload_dict(self):
         """ Reset `dictionary` to values from bound local file """
-        self.config["root"]['tobereset'] = False
-        self.config.reload()
+        self.language_object["root"]['tobereset'] = False
+        self.language_object.reload()
 
-        assert self.config.get('tobereset', True)
+        assert self.language_object.get('tobereset', True)
 
     def test_access_dashed_attribute(self):
-        assert self.config["root"]["dash-split"]["wow-man"] == default_local_file_dict["root"]["dash-split"]["wow-man"]
+        assert self.language_object["root"]["dash-split"]["wow-man"] == default_local_file_dict["root"]["dash-split"]["wow-man"]
 
     def test_add_dashed_attribute(self):
-        self.config["root"]["dash-split"]["second-dashed"] = "ok guys, bye"
-        self.config.commit()
+        self.language_object["root"]["dash-split"]["second-dashed"] = "ok guys, bye"
+        self.language_object.commit()
 
-        assert self.config["root"]["dash-split"]["second-dashed"] == \
-               self.config.read_file_as_dict()["root"]["dash-split"]["second-dashed"]
+        assert self.language_object["root"]["dash-split"]["second-dashed"] == \
+               self.language_object.read_file_as_dict()["root"]["dash-split"]["second-dashed"]
 
     def test_add_one_attribute(self):
         """
@@ -128,10 +128,10 @@ class BaseConfigTest():
         """
         key_name = "Masvingo"
 
-        self.config["root"]['city'] = key_name
-        self.config.commit()
+        self.language_object["root"]['city'] = key_name
+        self.language_object.commit()
 
-        assert self.config["root"]['city'] == self.config.read_file_as_dict()["root"]["city"]
+        assert self.language_object["root"]['city'] == self.language_object.read_file_as_dict()["root"]["city"]
 
     def test_add_nested_dicts(self):
         """
@@ -144,10 +144,10 @@ class BaseConfigTest():
             }
         }
 
-        self.config["root"]['items'] = dct
-        self.config.commit()
+        self.language_object["root"]['items'] = dct
+        self.language_object.commit()
 
-        assert int(self.config["root"]['items']['sugar']['carbon']) == int(self.config.read_file_as_dict()["root"]["items"]["sugar"]["carbon"])
+        assert int(self.language_object["root"]['items']['sugar']['carbon']) == int(self.language_object.read_file_as_dict()["root"]["items"]["sugar"]["carbon"])
 
     def test_add_multiple_dicts_with_update(self):
         """
@@ -163,20 +163,20 @@ class BaseConfigTest():
             }
         }
 
-        self.config["root"]['users'] = dct
-        self.config.commit()
+        self.language_object["root"]['users'] = dct
+        self.language_object.commit()
 
-        file_dict = self.config.read_file_as_dict()["root"]["users"]
+        file_dict = self.language_object.read_file_as_dict()["root"]["users"]
 
         print(file_dict["jerre_1829"]["status"])
-        print(self.config["root"]['users']['jerre_1829']['status'])
+        print(self.language_object["root"]['users']['jerre_1829']['status'])
 
-        assert int(self.config["root"]['users']['jerre_1829']['status']) == int(file_dict["jerre_1829"]["status"]) and \
-               int(self.config["root"]['users']['wolfrm4882']['status']) == int(file_dict["wolfrm4882"]["status"])
+        assert int(self.language_object["root"]['users']['jerre_1829']['status']) == int(file_dict["jerre_1829"]["status"]) and \
+               int(self.language_object["root"]['users']['wolfrm4882']['status']) == int(file_dict["wolfrm4882"]["status"])
 
     def test_modify_default_dict(self):
         """Modify the default dictionary, reset `dictionary` to defaults and assert values"""
-        self.config.dictionary_default["root"].update({"new_values": [256, 45, 154]})
-        self.config.reset_to_defaults()
+        self.language_object.dictionary_default["root"].update({"new_values": [256, 45, 154]})
+        self.language_object.reset_to_defaults()
 
-        assert self.config["root"].get('new_values', None) == self.config.dictionary_default["root"]['new_values']
+        assert self.language_object["root"].get('new_values', None) == self.language_object.dictionary_default["root"]['new_values']