+++ interform/tests/test_json.py
@@ -4,25 +4,25 @@ import pytest
 
 from interform import JSON_Format
 
-from .core import BaseConfigTest, local_file_path,\
+from .core import BaseLangTest, local_file_path,\
     default_local_file_dict, default_cfg_path
 
 
 @pytest.mark.json
-class Test_Create_DefDict(BaseConfigTest):
+class Test_Create_DefDict(BaseLangTest):
     """
     Create local file on `local_file_path`
     and get default config values from `default_local_file_dict`
     """
     def setup(self):
-        self.config = JSON_Format(
+        self.language_object = JSON_Format(
             local_file_path,
             default_local_file_dict
         )
 
 
 @pytest.mark.json
-class Test_Create_DefPath(BaseConfigTest):
+class Test_Create_DefPath(BaseLangTest):
     """
     Read existing file on `local_file_path`
     and get default config values from file on `default_cfg_path`
@@ -31,7 +31,7 @@ class Test_Create_DefPath(BaseConfigTest):
         with open(default_cfg_path, 'w') as f:
             json.dump(default_local_file_dict, f)
 
-        self.config = JSON_Format(
+        self.language_object = JSON_Format(
             local_file_path,
             default_cfg_path
         )